backend/pyproject.toml
[project]
name = "shadow-po-ai-backend"
version = "0.1.0"
description = "FastAPI backend for Shadow PO AI"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "bcrypt>=4.0.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.5",
  "openai>=1.57.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.10.1",
  "pydantic-settings>=2.6.1",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "uvicorn[standard]>=0.32.1"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
